requirements.txt
Flask==2.2.5
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.2
Flask-Bcrypt==1.0.1
qrcode==7.4.2
pandas==2.0.3
openpyxl==3.1.2
mysqlclient==2.2.7
Werkzeug==2.2.3
python-dotenv==1.0.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
gspread==5.12.4
oauth2client==4.1.3
